Fix c_long definition

+++ src/lib.rs
@@ -37,8 +37,8 @@ pub mod ctypes {
     pub type c_ushort = u16;
     pub type c_int = i32;
     pub type c_uint = u32;
-    pub type c_long = i32;
-    pub type c_ulong = u32;
+    pub type c_long = i64;
+    pub type c_ulong = u64;
     pub type c_longlong = i64;
     pub type c_ulonglong = u64;
     pub type c_float = f32;
